requirements.txt
# Core Framework
streamlit==1.28.0

# Data Processing
pandas==2.1.0
numpy==1.24.3
pillow==10.0.0

# Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0
opencv-python==4.8.0.76

# Geospatial Libraries
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0

# Visualization
plotly==5.16.1
matplotlib==3.7.2
seaborn==0.12.2
folium==0.14.0
streamlit-folium==0.13.0

# Utilities
kaggle==1.5.16
python-dotenv==1.0.0
tqdm==4.66.1
albumentations==1.3.1

# Optional: Deep Learning Enhancements
# torch==2.0.1
# torchvision==0.15.2
